Updated github CI action to use cargo cov instead of tarpaulin. Make make_release.sh use a version param instead of hardcoding

+++ tools/make_release.sh
@@ -1,5 +1,13 @@
 #!/usr/bin/env bash
-echo "Making release..."
+
+# Must be run with the version number as the only param. e.g.
+# ./make_release.sh 0.6.0
+if [[ $# -eq 0 ]] ; then
+    echo 'You must provide the release version number. E.g. ./make_release.sh 0.6.0'
+    exit 1
+fi
+
+echo "Making release... $1"
 echo "Cleaning up previous release files..."
 rm -rf ../target
 rm -rf release
@@ -18,5 +26,5 @@ cp ../CHANGELOG.md release/.
 cp ../include/mwalib.h release/include/.
 cd release
 echo "Taring files..."
-tar -czvf libmwalib-0.5.0-linux_x86_64.tar.gz lib/libmwalib.a lib/libmwalib.so include/mwalib.h LICENSE LICENSE-cfitsio CHANGELOG.md
+tar -czvf libmwalib-$1-linux_x86_64.tar.gz lib/libmwalib.a lib/libmwalib.so include/mwalib.h LICENSE LICENSE-cfitsio CHANGELOG.md
 echo "Release complete!"
